requirements.txt
Pillow>=8.1.1
PyYaml>=6.0.1
attrs>=15.0.0
numpy>=1.21.6
toml>=0.10.2
